refactor: 리뷰 점수 로직 BigDecimal 적용, 매너 점수 총점 및 기본점수 변경

+++ src/main/java/com/ddobang/backend/domain/member/dto/response/BasicProfileResponse.java
@@ -1,5 +1,7 @@
 package com.ddobang.backend.domain.member.dto.response;
 
+import java.math.BigDecimal;
+
 import com.ddobang.backend.domain.member.entity.Gender;
 import com.ddobang.backend.domain.member.entity.Member;
 
@@ -8,7 +10,7 @@ public record BasicProfileResponse(
 	Gender gender,
 	String introduction,
 	String profilePictureUrl,
-	int mannerScore
+	BigDecimal mannerScore
 ) {
 	public static BasicProfileResponse of(Member member) {
 		return new BasicProfileResponse(
